requirements.txt
Flask
psycopg2-binary
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
